pyproject.toml
[project]
name = "qnagpt"
version = "0.0.0"
description = "qnagpt is an AI based web app to generate questions from given pdfs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.9",
    "fastapi[standard]>=0.115.12",
    "mistralai>=1.7.0",
    "python-dotenv>=1.1.0",
]
